support custom assertions per test case
