+++ test/TestCases.test.ts
@@ -3,8 +3,7 @@ import * as fs from "fs-extra";
 import webpack = require("webpack");
 import MemoryFs = require("memory-fs");
 import pDefer = require("p-defer");
-import { normalizeStats } from "./_util/stats";
-import { satisfiesVersionRequirements } from "./_util/satisfiesVersionRequirements";
+import { createAssertExpectation, satisfiesVersionRequirements } from "./_util/testHelper";
 
 const testCasesPath = path.join(__dirname, "testCases");
 const tests = fs.readdirSync(testCasesPath).filter(dir => fs.statSync(path.join(testCasesPath, dir)).isDirectory());
@@ -23,6 +22,7 @@ describe("TestCases", () => {
     const tmpTestPath = path.join(tmpPath, testName);
     const webpackConfigPath = path.join(tmpTestPath, "webpack.config.ts");
     const skipTest = !satisfiesVersionRequirements(path.join(testPath, "versions.json"));
+    const assertExpectation = createAssertExpectation(path.join(testPath, "expectation.ts"));
 
     (skipTest ? it.skip : it)(testName, async () => {
       await fs.copy(testPath, tmpTestPath);
@@ -57,8 +57,7 @@ describe("TestCases", () => {
           return deferred.reject(err);
         }
 
-        const normalizedStats = normalizeStats(stats);
-        expect(normalizedStats).toMatchSnapshot();
+        assertExpectation(stats, 0);
 
         process.nextTick(deferred.resolve);
       });
